refactor: divide useNearestCourse hook
