+++ src/app/hooks/use-nearest-course/constants.ts
@@ -1,0 +1,1 @@
+export const BUFFER_PERIOD = 14;
