+++ src/app/hooks/use-nearest-course/index.ts
@@ -1,1 +1,2 @@
 export { useNearestCourse } from './use-nearest-course';
+export { finedNearestCourse } from './utils/fined-nearest-course';
