+++ src/app/hooks/use-nearest-course/types.ts
@@ -1,0 +1,1 @@
+export type Day = number;
