+++ src/app/hooks/use-nearest-course/utils/fined-nearest-course.ts
@@ -1,0 +1,11 @@
+import { BUFFER_PERIOD } from "../constants";
+import { Day } from "../types";
+import { chooseNearestCourse } from "./choose-nearest-course";
+import { definePrevNext } from "./define-prev-next";
+
+export function finedNearestCourse(courses: object[], bufferPeriod: Day = BUFFER_PERIOD) {
+  const { prevCourse, nextCourse } = definePrevNext(courses);  
+  const nearestCourse = chooseNearestCourse({prevCourse, nextCourse, bufferPeriod});
+  
+  return nearestCourse;
+}
